removed firm.co from list of tlds
Aparantly this is a valid domain rather than a 3rd level and a silly customer would like to register it.
Looking at the .co wiki page (http://en.wikipedia.org/wiki/.co) it may be that a few of the others you have listed in data/effective_tld_names.txt are also not real 3rd level domains.